fix git
